Move Field to sep file
